requirements.txt
gidgethub==4.2.0
discord.py==1.5.1
aiohttp==3.7.3